apollo_dashboard: split alerts to different modules (#8348)
