+++ crates/apollo_dashboard/src/alert_scenarios/mod.rs
@@ -1,0 +1,11 @@
+pub mod block_production_delay;
+pub mod block_production_halt;
+pub mod infra_alerts;
+pub mod l1_gas_prices;
+pub mod l1_handlers;
+pub mod mempool_size;
+pub mod preconfirmed;
+pub mod sync_halt;
+pub mod tps;
+pub mod transaction_delays;
+pub mod transaction_failures;
