+++ crates/apollo_dashboard/src/alert_scenarios/transaction_failures.rs
@@ -1,0 +1,181 @@
+use apollo_http_server::metrics::{
+    ADDED_TRANSACTIONS_DEPRECATED_ERROR,
+    ADDED_TRANSACTIONS_FAILURE,
+    ADDED_TRANSACTIONS_INTERNAL_ERROR,
+    ADDED_TRANSACTIONS_TOTAL,
+};
+use apollo_mempool::metrics::{MEMPOOL_TRANSACTIONS_DROPPED, MEMPOOL_TRANSACTIONS_RECEIVED};
+
+use crate::alerts::{
+    Alert,
+    AlertComparisonOp,
+    AlertCondition,
+    AlertEnvFiltering,
+    AlertGroup,
+    AlertLogicalOp,
+    AlertSeverity,
+};
+
+const PENDING_DURATION_DEFAULT: &str = "30s";
+const EVALUATION_INTERVAL_SEC_DEFAULT: u64 = 30;
+
+// TODO(guy.f): consider uniting with regular tx failure rate.
+// TODO(guyf.f): Change threshold to 0.05 after mainnet launch.
+pub(crate) fn get_http_server_high_deprecated_transaction_failure_ratio() -> Alert {
+    Alert::new(
+        "http_server_high_deprecated_transaction_failure_ratio",
+        "http server high deprecated transaction failure ratio",
+        AlertGroup::HttpServer,
+        format!(
+            "increase({}[1h]) / clamp_min(increase({}[1h]), 1)",
+            ADDED_TRANSACTIONS_DEPRECATED_ERROR.get_name_with_filter(),
+            ADDED_TRANSACTIONS_TOTAL.get_name_with_filter()
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::GreaterThan,
+            comparison_value: 0.1,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        AlertSeverity::Informational,
+        AlertEnvFiltering::All,
+    )
+}
+
+fn get_http_server_high_transaction_failure_ratio(
+    alert_env_filtering: AlertEnvFiltering,
+    alert_severity: AlertSeverity,
+) -> Alert {
+    Alert::new(
+        "http_server_high_transaction_failure_ratio",
+        "http server high transaction failure ratio",
+        AlertGroup::HttpServer,
+        format!(
+            "(increase({}[1h]) - increase({}[1h])) / clamp_min(increase({}[1h]), 1)",
+            ADDED_TRANSACTIONS_FAILURE.get_name_with_filter(),
+            ADDED_TRANSACTIONS_DEPRECATED_ERROR.get_name_with_filter(),
+            ADDED_TRANSACTIONS_TOTAL.get_name_with_filter()
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::GreaterThan,
+            comparison_value: 0.2,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        alert_severity,
+        alert_env_filtering,
+    )
+}
+
+pub(crate) fn get_http_server_high_transaction_failure_ratio_vec() -> Vec<Alert> {
+    vec![
+        get_http_server_high_transaction_failure_ratio(
+            AlertEnvFiltering::MainnetStyleAlerts,
+            AlertSeverity::Regular,
+        ),
+        get_http_server_high_transaction_failure_ratio(
+            AlertEnvFiltering::TestnetStyleAlerts,
+            AlertSeverity::WorkingHours,
+        ),
+    ]
+}
+
+fn get_http_server_internal_error_ratio(
+    alert_env_filtering: AlertEnvFiltering,
+    alert_severity: AlertSeverity,
+) -> Alert {
+    Alert::new(
+        "http_server_internal_error_ratio",
+        "http server internal error ratio",
+        AlertGroup::HttpServer,
+        format!(
+            "increase({}[1h]) / clamp_min(increase({}[1h]), 1)",
+            ADDED_TRANSACTIONS_INTERNAL_ERROR.get_name_with_filter(),
+            ADDED_TRANSACTIONS_TOTAL.get_name_with_filter()
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::GreaterThan,
+            comparison_value: 0.01,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        alert_severity,
+        alert_env_filtering,
+    )
+}
+
+pub(crate) fn get_http_server_internal_error_ratio_vec() -> Vec<Alert> {
+    vec![
+        get_http_server_internal_error_ratio(
+            AlertEnvFiltering::MainnetStyleAlerts,
+            AlertSeverity::Regular,
+        ),
+        get_http_server_internal_error_ratio(
+            AlertEnvFiltering::TestnetStyleAlerts,
+            AlertSeverity::WorkingHours,
+        ),
+    ]
+}
+
+fn get_mempool_transaction_drop_ratio(
+    alert_env_filtering: AlertEnvFiltering,
+    alert_severity: AlertSeverity,
+) -> Alert {
+    Alert::new(
+        "mempool_transaction_drop_ratio",
+        "Mempool transaction drop ratio",
+        AlertGroup::Mempool,
+        format!(
+            "increase({}[10m]) / clamp_min(increase({}[10m]), 1)",
+            MEMPOOL_TRANSACTIONS_DROPPED.get_name_with_filter(),
+            MEMPOOL_TRANSACTIONS_RECEIVED.get_name_with_filter(),
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::GreaterThan,
+            // TODO(leo): Decide on the final ratio and who should be alerted.
+            comparison_value: 0.2,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        alert_severity,
+        alert_env_filtering,
+    )
+}
+
+pub(crate) fn get_mempool_transaction_drop_ratio_vec() -> Vec<Alert> {
+    vec![
+        get_mempool_transaction_drop_ratio(
+            AlertEnvFiltering::MainnetStyleAlerts,
+            AlertSeverity::DayOnly,
+        ),
+        get_mempool_transaction_drop_ratio(
+            AlertEnvFiltering::TestnetStyleAlerts,
+            AlertSeverity::WorkingHours,
+        ),
+    ]
+}
+
+pub(crate) fn get_http_server_internal_error_once() -> Alert {
+    Alert::new(
+        "http_server_internal_error_once",
+        "http server internal error once",
+        AlertGroup::HttpServer,
+        format!(
+            "increase({}[20m]) or vector(0)",
+            ADDED_TRANSACTIONS_INTERNAL_ERROR.get_name_with_filter()
+        ),
+        vec![AlertCondition {
+            comparison_op: AlertComparisonOp::GreaterThan,
+            comparison_value: 0.0,
+            logical_op: AlertLogicalOp::And,
+        }],
+        PENDING_DURATION_DEFAULT,
+        EVALUATION_INTERVAL_SEC_DEFAULT,
+        AlertSeverity::WorkingHours,
+        AlertEnvFiltering::All,
+    )
+}
