+++ crates/apollo_dashboard/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod alert_definitions;
+pub mod alert_scenarios;
 pub mod alerts;
 mod dashboard;
 pub mod dashboard_definitions;
